asl-find-sound should pickup the same audio interfaces
The original "asl-find-sound" walked the USB devices on the system looking for
sound/audio interfaces based on the "manufacturer" string.  The scrit was later
updated to also pickup interfaces that did not have a matching string bug did
have a CM108 vendor/product ID.  This change updates the script to find the same
interfaces as SimpleUSB and USBRadio.